refactor(core): delete TlsTransport now that every bridge runs on WireMeshTransport
TlsTransport was agent-comms' own hand-rolled MeshTransport implementation
over X.509 certificates and newline-delimited JSON framing -- the
substrate every bridge ran on before the P2 migration. With all six
bridges cut over to WireMeshTransport (over @exadev/wire-mesh-core) and
its own connection-approval quarantine gate closing the security gap that
substrate swap introduced, nothing constructs a TlsTransport any more.

test-transport.ts's wireTestTransport and wireWireMeshTestTransport
wired two different transports onto a test MeshStore for exactly this
migration period; with only one transport left, they collapse into one
function (kept the wireTestTransport name, the one every existing test
already called). Every test that built its own peer directly against
TlsTransport (broadcast-window, downtime-replay.integration,
identity-restart) now does the same against WireMeshTransport, with
fingerprint-based identity comparisons swapped for deviceId, matching the
canonical identity createBridgeMesh already uses in production.

listener-policy.integration.test.ts's one test that spoke TlsTransport's
own wire format directly (a raw tls.connect() writing newline-delimited
JSON) is rewritten against a real WireMeshTransport client session
instead, and no longer reaches into MeshStore's internals through an
unsafe cast to observe the result -- it constructs its own transport
directly with a hand-rolled TransportEvents object instead, the same
pattern become-coordinator-actual-port.integration.test.ts already used.

peer-id-verification.integration.test.ts is deleted outright rather than
migrated: the vulnerability class it guarded against (a peer's
self-reported peerId not matching the certificate it authenticated with)
cannot occur under WireMeshTransport by construction, since peer identity
is never self-reported in the first place -- it's read directly from the
TLS-authenticated connection's own peerDeviceId.

wire-mesh-transport.integration.test.ts and
wire-mesh-transport-approval.integration.test.ts were written during the
migration specifically to prove WireMeshTransport reached parity with
TlsTransport's own behaviour. Now that wireTestTransport wires
WireMeshTransport for every test, they duplicate mesh-e2e.integration.test.ts
and approval.integration.test.ts/listener-policy.integration.test.ts
scenario-for-scenario. Deleted, after moving the one genuinely unique
test in the approval file over: the quarantine-gate regression test
proving a forged state_update from an unapproved connection is refused.

Fixed along the way: approval.integration.test.ts's own "mesh_pending
lists pending connections" test shut down storeB immediately once its
own assertions were satisfied, without waiting for storeB's own
fire-and-forget connectToRemote continuation (the mesh-formation
handshake) to actually settle. Under WireMeshTransport's real round-trip
timing this let the test's own shutdown race ahead of that continuation,
which then tried to track a session into an already-shut-down transport
and never closed it -- the same convergence wait "accept establishes the
peer connection" already uses fixes it.
