+++ src/test/downtime-replay.integration.test.ts
@@ -8,8 +8,9 @@ import * as assert from "node:assert/strict";
 import * as fs from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
+import { deviceIdToHex } from "@exadev/wire-mesh-core/domain/device-id";
 import { MeshStore } from "../core/mesh-store.js";
-import { TlsTransport } from "../core/tls-transport.js";
+import { WireMeshTransport } from "../core/wire-mesh-transport.js";
 import { generateIdentity } from "../core/identity.js";
 import {
   loadOrCreateIdentity,
@@ -27,11 +28,11 @@ interface Peer {
   deliveries: DeliveryEvent[];
 }
 
-/** A peer wired like a real bridge: TLS transport, fingerprint peer ID. */
+/** A peer wired like a real bridge: WireMeshTransport, device-id peer ID. */
 function makePeer(identity: PeerIdentity): Peer {
   const store = new MeshStore(TEST_PORT);
-  store.peerId = identity.fingerprint;
-  store.setTransport(new TlsTransport(store.events, identity));
+  store.peerId = deviceIdToHex(Uint8Array.from(identity.deviceId));
+  store.setTransport(new WireMeshTransport(store.events, identity));
   const deliveries: DeliveryEvent[] = [];
   return { store, deliveries };
 }
@@ -109,7 +110,10 @@ async function main(): Promise<void> {
 
   // B restarts in the same slot: same identity, same agent ID.
   const identityB2 = loadOrCreateIdentity(slot);
-  assert.equal(identityB2.fingerprint, identityB.fingerprint);
+  assert.equal(
+    deviceIdToHex(Uint8Array.from(identityB2.deviceId)),
+    deviceIdToHex(Uint8Array.from(identityB.deviceId)),
+  );
   const b2 = makePeer(identityB2);
   b2.store.onDelivery = (_id, ev) => {
     b2.deliveries.push(ev);
